+++ packages/interactivity-router/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   Add `state.navigating` and `state.initiator` to the `core/router` store, plus an `initiator` option for `navigate()`, giving extenders a supported, reactive way to observe client-side navigations and know what started them. ([#TBD](https://github.com/WordPress/gutenberg/pull/TBD))
+
 ## 2.55.0 (2026-09-10)
 
 ## 2.54.0 (2026-08-26)